+++ src/js/interfaces/VendingMachine.interface.ts
@@ -1,0 +1,14 @@
+interface Coin {
+  coin10: number;
+  coin50: number;
+  coin100: number;
+  coin500: number;
+}
+
+interface Product {
+  name: string;
+  amount: number;
+  price: number;
+}
+
+export { Coin, Product };
